requirements.txt
langgraph
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-openai
streamlit
python-docx
opencv-python
yt-dlp
moviepy
google-generativeai
python-dotenv
playwright
mcp
uvicorn
fastapi
pyparsing<3.0.0